♻️ refactor: extract user type
